pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idle_detector"
version = "0.0.1"
authors = [
{ name = "Yousef Abuzahrieh", email = "[yousefzahrieh17@gmail.com](mailto:yousefzahrieh17@gmail.com)" }
]
description = "IdleDetector is a macOS-only Python library and CLI for monitoring user idle time and sending stage alerts via terminal-notifier."
requires-python = ">=3.13"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
"Intended Audience :: Developers",
"License :: OSI Approved :: Apache Software License",
"Operating System :: OS Independent",
"Programming Language :: Python :: 3",
"Programming Language :: Python :: 3 :: Only",
"Topic :: Utilities",
"Topic :: System :: Monitoring"
]

dependencies = [
    "platformdirs",
    "pyobjc-framework-Quartz",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff"
]

[project.urls]
Homepage = "[https://github.com/yousefabuz17/idleDetector](https://github.com/yousefabuz17/idleDetector)"
Download = "[https://github.com/yousefabuz17/idleDetector.git](https://github.com/yousefabuz17/idleDetector.git)"

[project.scripts]
idle-detector = "idle_detector.cli:cli_parser"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.uv.sources]
idle_detector = { path = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [
    "build",
    "dist",
    ".venv",
    ".uv",
    ".git",
    "**/**pycache**"
]
include = ["src/idle_detector"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
fixable = ["ALL"]
